@@ -1,102 +1,216 @@
 <Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<OutputType>WinExe</OutputType>
-		<TargetFramework>net7.0-windows10.0.19041.0</TargetFramework>
-		<TargetPlatformMinVersion>10.0.18362.0</TargetPlatformMinVersion>
-		<RootNamespace>ProjectSBS.Windows</RootNamespace>
-		<ApplicationManifest>app.manifest</ApplicationManifest>
-		<Platforms>x86;x64;arm64</Platforms>
-		<RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
-		<PublishProfile>win10-$(Platform).pubxml</PublishProfile>
-		<UseWinUI>true</UseWinUI>
-		<EnableMsixTooling>true</EnableMsixTooling>
-		<!-- Uncomment for Unpackaged builds -->
-		<!-- <WindowsPackageType>None</WindowsPackageType> -->
-    	<!-- For Unpackaged builds, change version of SDK to previous stable release https://github.com/microsoft/WindowsAppSDK/issues/3591 -->
-	</PropertyGroup>
-	<PropertyGroup>
-		<!-- Bundles the WinAppSDK binaries (Uncomment for unpackaged builds) -->
-		<!-- <WindowsAppSDKSelfContained>true</WindowsAppSDKSelfContained> -->
-		<!-- <SelfContained>true</SelfContained> -->
-	</PropertyGroup>
+  <PropertyGroup>
+    <OutputType>WinExe</OutputType>
+    <TargetFramework>net8.0-windows10.0.22621.0</TargetFramework>
+    <RootNamespace>ProjectSBS.Windows</RootNamespace>
+    <ApplicationManifest>app.manifest</ApplicationManifest>
+    <Platforms>x86;x64;arm64</Platforms>
+    <RuntimeIdentifiers>win-x86;win-x64;win-arm64</RuntimeIdentifiers>
+    <PublishProfile>win-$(Platform).pubxml</PublishProfile>
+    <UseWinUI>true</UseWinUI>
+    <EnableMsixTooling>true</EnableMsixTooling>
+    <GenerateTemporaryStoreCertificate>True</GenerateTemporaryStoreCertificate>
+    <GenerateAppInstallerFile>False</GenerateAppInstallerFile>
+    <AppxPackageSigningEnabled>False</AppxPackageSigningEnabled>
+    <AppxPackageSigningTimestampDigestAlgorithm>SHA256</AppxPackageSigningTimestampDigestAlgorithm>
+    <AppxAutoIncrementPackageRevision>False</AppxAutoIncrementPackageRevision>
+    <AppxPackageDir>D:\Documents\Projects\xx\SBSRelease</AppxPackageDir>
+    <AppxSymbolPackageEnabled>True</AppxSymbolPackageEnabled>
+    <GenerateTestArtifacts>True</GenerateTestArtifacts>
+    <AppxBundle>Never</AppxBundle>
+    <AppxBundlePlatforms>x86|x64|arm64</AppxBundlePlatforms>
+    <HoursBetweenUpdateChecks>0</HoursBetweenUpdateChecks>
+    <SupportedOSPlatformVersion>10.0.19041.0</SupportedOSPlatformVersion>
+    <!--<PublishTrimmed>true</PublishTrimmed>-->
 
-	<ItemGroup>
-		<Content Include="Images\**" />
-		<Manifest Include="$(ApplicationManifest)" />
-	</ItemGroup>
+    <!--<ErrorOnDuplicatePublishOutputFiles>false</ErrorOnDuplicatePublishOutputFiles>-->
 
-	<ItemGroup>
-		<PackageReference Include="CommunityToolkit.Labs.WinUI.SettingsControls" />
-		<PackageReference Include="CommunityToolkit.WinUI.Converters" />
-		<PackageReference Include="CommunityToolkit.WinUI.Triggers" />
-		<PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
-		<PackageReference Include="Microsoft.AppCenter" />
-		<PackageReference Include="Microsoft.AppCenter.Analytics" />
-		<PackageReference Include="Microsoft.AppCenter.Crashes" />
-		<PackageReference Include="Uno.WinUI" />
-		<PackageReference Include="Microsoft.WindowsAppSDK" />
-		<PackageReference Include="Microsoft.Windows.SDK.BuildTools" />
-		<PackageReference Include="CommunityToolkit.Mvvm" />
-		<PackageReference Include="Uno.Extensions.Configuration" />
-		<PackageReference Include="Uno.Extensions.Http" />
-		<PackageReference Include="Uno.Extensions.Http.Refit" />
-		<PackageReference Include="Uno.Extensions.Logging.WinUI" />
-		<PackageReference Include="Uno.Extensions.Serialization.Http" />
-		<PackageReference Include="Uno.Extensions.Serialization.Refit" />
-		<PackageReference Include="Uno.Extensions.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Authentication.WinUI" />
-		<PackageReference Include="Microsoft.Identity.Client" />
-		<PackageReference Include="Uno.Extensions.Authentication.MSAL.WinUI" />
-		<PackageReference Include="Uno.Extensions.Hosting.WinUI" />
-		<PackageReference Include="Uno.Extensions.Localization.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.WinUI" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" />
-		<PackageReference Include="Uno.Core.Extensions.Logging.Singleton" />
-		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" />
-		<PackageReference Include="Microsoft.Graph" />
-		<PackageReference Include="CommunityToolkit.WinUI.Notifications" />
-		<PackageReference Include="WinUIEx" />
-	</ItemGroup>
+    <!-- Bundles the WinAppSDK binaries -->
+    <!--<WindowsAppSDKSelfContained>true</WindowsAppSDKSelfContained>-->
 
-	<ItemGroup>
-		<!--
-		If you encounter this error message:
+    <!-- <SelfContained>true</SelfContained> -->
+  </PropertyGroup>
 
-			error NETSDK1148: A referenced assembly was compiled using a newer version of Microsoft.Windows.SDK.NET.dll.
-			Please update to a newer .NET SDK in order to reference this assembly.
+  <ItemGroup>
+    <Content Include="Images\**" />
+    <Manifest Include="$(ApplicationManifest)" />
+  </ItemGroup>
 
-		This means that the two packages below must be aligned with the "build" version number of
-		the "Microsoft.Windows.SDK.BuildTools" package above, and the "revision" version number
-		must be the highest found in https://www.nuget.org/packages/Microsoft.Windows.SDK.NET.Ref.
-		-->
-	    <FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" RuntimeFrameworkVersion="10.0.22621.28" />
-		<FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" TargetingPackVersion="10.0.22621.28" /> 
-	</ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" />
+    <PackageReference Include="Uno.WinUI" />
+    <PackageReference Include="Microsoft.WindowsAppSDK" />
+    <PackageReference Include="Microsoft.Windows.SDK.BuildTools" />
+    <PackageReference Include="CommunityToolkit.Mvvm" />
+    <PackageReference Include="Uno.Extensions.Configuration" />
+    <PackageReference Include="Uno.Extensions.Http" />
+    <PackageReference Include="Uno.Extensions.Http.Refit" />
+    <PackageReference Include="Uno.Extensions.Logging.WinUI" />
+    <PackageReference Include="Uno.Extensions.Serialization.Http" />
+    <PackageReference Include="Uno.Extensions.Serialization.Refit" />
+    <PackageReference Include="Uno.Toolkit.WinUI" />
+    <PackageReference Include="Uno.Extensions.Hosting.WinUI" />
+    <PackageReference Include="Uno.Extensions.Localization.WinUI" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" />
+    <PackageReference Include="Uno.Core.Extensions.Logging.Singleton" />
+    <PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" />
 
-	<ItemGroup>
-		<ProjectReference Include="..\ProjectSBS\ProjectSBS.csproj" />
-		<ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
-	</ItemGroup>
+    <PackageReference Include="WinUIEx" />
+    <PackageReference Include="CommunityToolkit.WinUI.Notifications" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.Segmented" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.SettingsControls" />
+    <PackageReference Include="CommunityToolkit.WinUI.Converters" />
+    <PackageReference Include="CommunityToolkit.WinUI.Animations" />
+    <PackageReference Include="CommunityToolkit.WinUI.Triggers" />
+    <PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
+    <PackageReference Include="Microsoft.AppCenter" />
+    <PackageReference Include="Microsoft.AppCenter.Analytics" />
+    <PackageReference Include="Microsoft.AppCenter.Crashes" />
+    <PackageReference Include="Microsoft.Graph" />
+    <PackageReference Include="Microsoft.Identity.Client" />
+    <PackageReference Include="Uno.WinUI.MSAL" />
+  </ItemGroup>
 
-	<!-- 
-		Defining the "Msix" ProjectCapability here allows the Single-project MSIX Packaging
-		Tools extension to be activated for this project even if the Windows App SDK Nuget
-		package has not yet been restored.
-	-->
-	<ItemGroup Condition="'$(DisableMsixProjectCapabilityAddedByProject)'!='true' and '$(EnableMsixTooling)'=='true'">
-		<ProjectCapability Include="Msix" />
-	</ItemGroup>
+  <ItemGroup><!--
+    If you encounter this error message:
 
-	<!-- 
-		Defining the "HasPackageAndPublishMenuAddedByProject" property here allows the Solution 
-		Explorer "Package and Publish" context menu entry to be enabled for this project even if 
-		the Windows App SDK Nuget package has not yet been restored.
-	-->
-	<PropertyGroup Condition="'$(DisableHasPackageAndPublishMenuAddedByProject)'!='true' and '$(EnableMsixTooling)'=='true'">
-		<HasPackageAndPublishMenu>true</HasPackageAndPublishMenu>
-	</PropertyGroup>
+      error NETSDK1148: A referenced assembly was compiled using a newer version of Microsoft.Windows.SDK.NET.dll.
+      Please update to a newer .NET SDK in order to reference this assembly.
 
-	<Import Project="..\ProjectSBS.Base\base.props" />
+    This means that the two packages below must be aligned with the "build" version number of
+    the "Microsoft.Windows.SDK.BuildTools" package above, and the "revision" version number
+    must be the highest found in https://www.nuget.org/packages/Microsoft.Windows.SDK.NET.Ref.
+    -->
+    <!-- <FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" RuntimeFrameworkVersion="10.0.22621.28" />
+    <FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" TargetingPackVersion="10.0.22621.28" /> -->
+  
+    <None Remove="AppIcon.scale-100.png" />
+    <None Remove="AppIcon.scale-125.png" />
+    <None Remove="AppIcon.scale-150.png" />
+    <None Remove="AppIcon.scale-200.png" />
+    <None Remove="AppIcon.scale-400.png" />
+    <None Remove="AppIcon.targetsize-16.png" />
+    <None Remove="AppIcon.targetsize-16_altform-lightunplated.png" />
+    <None Remove="AppIcon.targetsize-16_altform-unplated.png" />
+    <None Remove="AppIcon.targetsize-24.png" />
+    <None Remove="AppIcon.targetsize-24_altform-lightunplated.png" />
+    <None Remove="AppIcon.targetsize-24_altform-unplated.png" />
+    <None Remove="AppIcon.targetsize-256.png" />
+    <None Remove="AppIcon.targetsize-256_altform-lightunplated.png" />
+    <None Remove="AppIcon.targetsize-256_altform-unplated.png" />
+    <None Remove="AppIcon.targetsize-32.png" />
+    <None Remove="AppIcon.targetsize-32_altform-lightunplated.png" />
+    <None Remove="AppIcon.targetsize-32_altform-unplated.png" />
+    <None Remove="AppIcon.targetsize-48.png" />
+    <None Remove="AppIcon.targetsize-48_altform-lightunplated.png" />
+    <None Remove="AppIcon.targetsize-48_altform-unplated.png" />
+    <None Remove="LargeTile.scale-100.png" />
+    <None Remove="LargeTile.scale-125.png" />
+    <None Remove="LargeTile.scale-150.png" />
+    <None Remove="LargeTile.scale-200.png" />
+    <None Remove="LargeTile.scale-400.png" />
+    <None Remove="MediumTile.scale-100.png" />
+    <None Remove="MediumTile.scale-125.png" />
+    <None Remove="MediumTile.scale-150.png" />
+    <None Remove="MediumTile.scale-200.png" />
+    <None Remove="MediumTile.scale-400.png" />
+    <None Remove="PackageLogo.scale-100.png" />
+    <None Remove="PackageLogo.scale-125.png" />
+    <None Remove="PackageLogo.scale-150.png" />
+    <None Remove="PackageLogo.scale-200.png" />
+    <None Remove="PackageLogo.scale-400.png" />
+    <None Remove="SmallTile.scale-100.png" />
+    <None Remove="SmallTile.scale-125.png" />
+    <None Remove="SmallTile.scale-150.png" />
+    <None Remove="SmallTile.scale-200.png" />
+    <None Remove="SmallTile.scale-400.png" />
+    <None Remove="SplashScreen.scale-100.png" />
+    <None Remove="SplashScreen.scale-125.png" />
+    <None Remove="SplashScreen.scale-150.png" />
+    <None Remove="SplashScreen.scale-200.png" />
+    <None Remove="SplashScreen.scale-400.png" />
+    <None Remove="WideTile.scale-100.png" />
+    <None Remove="WideTile.scale-125.png" />
+    <None Remove="WideTile.scale-150.png" />
+    <None Remove="WideTile.scale-200.png" />
+    <None Remove="WideTile.scale-400.png" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="AppIcon.scale-100.png" />
+    <Content Include="AppIcon.scale-125.png" />
+    <Content Include="AppIcon.scale-150.png" />
+    <Content Include="AppIcon.scale-200.png" />
+    <Content Include="AppIcon.scale-400.png" />
+    <Content Include="AppIcon.targetsize-16.png" />
+    <Content Include="AppIcon.targetsize-16_altform-lightunplated.png" />
+    <Content Include="AppIcon.targetsize-16_altform-unplated.png" />
+    <Content Include="AppIcon.targetsize-24.png" />
+    <Content Include="AppIcon.targetsize-24_altform-lightunplated.png" />
+    <Content Include="AppIcon.targetsize-24_altform-unplated.png" />
+    <Content Include="AppIcon.targetsize-256.png" />
+    <Content Include="AppIcon.targetsize-256_altform-lightunplated.png" />
+    <Content Include="AppIcon.targetsize-256_altform-unplated.png" />
+    <Content Include="AppIcon.targetsize-32.png" />
+    <Content Include="AppIcon.targetsize-32_altform-lightunplated.png" />
+    <Content Include="AppIcon.targetsize-32_altform-unplated.png" />
+    <Content Include="AppIcon.targetsize-48.png" />
+    <Content Include="AppIcon.targetsize-48_altform-lightunplated.png" />
+    <Content Include="AppIcon.targetsize-48_altform-unplated.png" />
+    <Content Include="LargeTile.scale-100.png" />
+    <Content Include="LargeTile.scale-125.png" />
+    <Content Include="LargeTile.scale-150.png" />
+    <Content Include="LargeTile.scale-200.png" />
+    <Content Include="LargeTile.scale-400.png" />
+    <Content Include="MediumTile.scale-100.png" />
+    <Content Include="MediumTile.scale-125.png" />
+    <Content Include="MediumTile.scale-150.png" />
+    <Content Include="MediumTile.scale-200.png" />
+    <Content Include="MediumTile.scale-400.png" />
+    <Content Include="PackageLogo.scale-100.png" />
+    <Content Include="PackageLogo.scale-125.png" />
+    <Content Include="PackageLogo.scale-150.png" />
+    <Content Include="PackageLogo.scale-200.png" />
+    <Content Include="PackageLogo.scale-400.png" />
+    <Content Include="SmallTile.scale-100.png" />
+    <Content Include="SmallTile.scale-125.png" />
+    <Content Include="SmallTile.scale-150.png" />
+    <Content Include="SmallTile.scale-200.png" />
+    <Content Include="SmallTile.scale-400.png" />
+    <Content Include="SplashScreen.scale-100.png" />
+    <Content Include="SplashScreen.scale-125.png" />
+    <Content Include="SplashScreen.scale-150.png" />
+    <Content Include="SplashScreen.scale-200.png" />
+    <Content Include="SplashScreen.scale-400.png" />
+    <Content Include="WideTile.scale-100.png" />
+    <Content Include="WideTile.scale-125.png" />
+    <Content Include="WideTile.scale-150.png" />
+    <Content Include="WideTile.scale-200.png" />
+    <Content Include="WideTile.scale-400.png" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ProjectSBS\ProjectSBS.csproj" />
+    <ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
+  </ItemGroup>
+
+  <!--
+    Defining the "Msix" ProjectCapability here allows the Single-project MSIX Packaging
+    Tools extension to be activated for this project even if the Windows App SDK Nuget
+    package has not yet been restored.
+  -->
+  <ItemGroup Condition="'$(DisableMsixProjectCapabilityAddedByProject)'!='true' and '$(EnableMsixTooling)'=='true'">
+    <ProjectCapability Include="Msix" />
+  </ItemGroup>
+
+  <!--
+    Defining the "HasPackageAndPublishMenuAddedByProject" property here allows the Solution
+    Explorer "Package and Publish" context menu entry to be enabled for this project even if
+    the Windows App SDK Nuget package has not yet been restored.
+  -->
+  <PropertyGroup Condition="'$(DisableHasPackageAndPublishMenuAddedByProject)'!='true' and '$(EnableMsixTooling)'=='true'">
+    <HasPackageAndPublishMenu>true</HasPackageAndPublishMenu>
+  </PropertyGroup>
+
+  <Import Project="..\ProjectSBS.Base\base.props" />
 </Project>
