@@ -1,118 +1,153 @@
 <Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows')) or '$(EnableWindowsTargeting)' == 'true'">$(TargetFrameworks);net7.0-windows10.0.19041</TargetFrameworks>
-		<TargetFrameworks>$(TargetFrameworks);net7.0;net7.0-ios;net7.0-android;net7.0-maccatalyst</TargetFrameworks>
+  <PropertyGroup>
+    <TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows')) or '$(EnableWindowsTargeting)' == 'true'">$(TargetFrameworks);net8.0-windows10.0.19041</TargetFrameworks>
+    <TargetFrameworks>$(TargetFrameworks);net8.0;net8.0-android;net8.0-ios;net8.0-maccatalyst</TargetFrameworks>
+    <TargetFrameworks Condition="'$(OverrideTargetFrameworks)'!=''">$(OverrideTargetFrameworks)</TargetFrameworks>
 
-		<!-- Ensures the .xr.xml files are generated in a proper layout folder -->
-		<GenerateLibraryLayout>true</GenerateLibraryLayout>
-	</PropertyGroup>
+    <!-- Ensures the .xr.xml files are generated in a proper layout folder -->
+    <GenerateLibraryLayout>true</GenerateLibraryLayout>
+  </PropertyGroup>
 
-	<ItemGroup>
-		<PackageReference Include="CommunityToolkit.Labs.WinUI.SettingsControls" />
-		<PackageReference Include="CommunityToolkit.WinUI.Converters" />
-		<PackageReference Include="CommunityToolkit.WinUI.Triggers" />
-		<PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
-		<PackageReference Include="Microsoft.AppCenter" />
-		<PackageReference Include="Microsoft.AppCenter.Analytics" />
-		<PackageReference Include="Microsoft.AppCenter.Crashes" />
-		<PackageReference Include="Uno.WinUI" />
-		<PackageReference Include="Uno.Resizetizer" />
-		<PackageReference Include="CommunityToolkit.Mvvm" />
-		<PackageReference Include="Uno.Extensions.Configuration" />
-		<PackageReference Include="Uno.Extensions.Http" />
-		<PackageReference Include="Uno.Extensions.Http.Refit" />
-		<PackageReference Include="Uno.Extensions.Logging.WinUI" />
-		<PackageReference Include="Uno.Extensions.Serialization.Http" />
-		<PackageReference Include="Uno.Extensions.Serialization.Refit" />
-		<PackageReference Include="Uno.Extensions.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Authentication.WinUI" />
-		<PackageReference Include="Microsoft.Identity.Client" />
-		<PackageReference Include="Uno.Extensions.Authentication.MSAL.WinUI" />
-		<PackageReference Include="Uno.Extensions.Hosting.WinUI" />
-		<PackageReference Include="Uno.Extensions.Localization.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.WinUI" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" />
-		<PackageReference Include="Microsoft.Graph" />
-		<PackageReference Include="CommunityToolkit.WinUI.Notifications" />
-		<PackageReference Include="WinUIEx" Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'windows'" />
-	</ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" />
+    <PackageReference Include="Uno.WinUI" />
+    <PackageReference Include="Uno.Resizetizer" />
+    <PackageReference Include="CommunityToolkit.Mvvm" />
+    <PackageReference Include="Uno.Extensions.Configuration" />
+    <PackageReference Include="Uno.Extensions.Http" />
+    <PackageReference Include="Uno.Extensions.Http.Refit" />
+    <PackageReference Include="Uno.Extensions.Logging.WinUI" />
+    <PackageReference Include="Uno.Extensions.Serialization.Http" />
+    <PackageReference Include="Uno.Extensions.Http.WinUI" />
+    <PackageReference Include="Uno.Extensions.Serialization.Refit" />
+    <PackageReference Include="Uno.Toolkit.WinUI" />
+    <PackageReference Include="Uno.Extensions.Core.WinUI" />
+    <PackageReference Include="Uno.Extensions.Hosting.WinUI" />
+    <PackageReference Include="Uno.Extensions.Localization.WinUI" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" />
 
-	<Choose>
-		<When Condition="$(IsWinAppSdk)">
-			<PropertyGroup>
-				<!--
-				If you encounter this error message:
+    <PackageReference Include="WinUIEx" Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'windows'" />
+    <PackageReference Include="CommunityToolkit.WinUI.Notifications" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.Segmented" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.SettingsControls" />
+    <PackageReference Include="CommunityToolkit.WinUI.Converters" />
+    <PackageReference Include="CommunityToolkit.WinUI.Animations" />
+    <PackageReference Include="CommunityToolkit.WinUI.Triggers" />
+    <PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
+    <PackageReference Include="Microsoft.AppCenter" />
+    <PackageReference Include="Microsoft.AppCenter.Analytics" />
+    <PackageReference Include="Microsoft.AppCenter.Crashes" />
+    <PackageReference Include="Microsoft.Graph" />
+    <PackageReference Include="Microsoft.Identity.Client" />
+    <PackageReference Include="Uno.WinUI.MSAL" />
+  </ItemGroup>
 
-					error NETSDK1148: A referenced assembly was compiled using a newer version of Microsoft.Windows.SDK.NET.dll.
-					Please update to a newer .NET SDK in order to reference this assembly.
 
-				This means that the two packages below must be aligned with the "build" version number of
-				the "Microsoft.Windows.SDK.BuildTools" package above, and the "revision" version number
-				must be the highest found in https://www.nuget.org/packages/Microsoft.Windows.SDK.NET.Ref.
-				-->
-			    <WindowsSdkPackageVersion>10.0.22621.28</WindowsSdkPackageVersion> 
-				<RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
-			</PropertyGroup>
-			<ItemGroup>
-				<PackageReference Include="Microsoft.WindowsAppSDK" />
-				<PackageReference Include="Microsoft.Windows.SDK.BuildTools" />
-			</ItemGroup>
-		</When>
-		<Otherwise>
-			<ItemGroup>
-				<Content Include="Assets\**" />
-				<Page Include="**\*.xaml" Exclude="bin\**\*.xaml;obj\**\*.xaml" />
-				<Compile Update="**\*.xaml.cs">
-					<DependentUpon>%(Filename)</DependentUpon>
-				</Compile>
-				<PRIResource Include="**\*.resw" />
-			</ItemGroup>
-		</Otherwise>
-	</Choose>
+  <Choose>
+    <When Condition="$(IsWinAppSdk)">
+      <PropertyGroup>
+        <!--
+        If you encounter this error message:
 
-	<ItemGroup>
-		<ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
-	</ItemGroup>
+          error NETSDK1148: A referenced assembly was compiled using a newer version of Microsoft.Windows.SDK.NET.dll.
+          Please update to a newer .NET SDK in order to reference this assembly.
 
-	<ItemGroup>
-		<UnoImage Include="Assets\**\*.svg" />
-		<EmbeddedResource Include="appsettings.json" />
-		<EmbeddedResource Include="appsettings.*.json" DependentUpon="appsettings.json" />
-		<UpToDateCheckInput Include="**\*.xaml" Exclude="bin\**\*.xaml;obj\**\*.xaml" />
-	</ItemGroup>
+        This means that the two packages below must be aligned with the "build" version number of
+        the "Microsoft.Windows.SDK.BuildTools" package above, and the "revision" version number
+        must be the highest found in https://www.nuget.org/packages/Microsoft.Windows.SDK.NET.Ref.
+        -->
+        <!-- <WindowsSdkPackageVersion>10.0.22621.28</WindowsSdkPackageVersion> -->
+      </PropertyGroup>
+      <ItemGroup>
+        <PackageReference Include="Microsoft.WindowsAppSDK" />
+        <PackageReference Include="Microsoft.Windows.SDK.BuildTools" />
+      </ItemGroup>
+    </When>
+    <Otherwise>
+      <ItemGroup>
+        <PackageReference Include="Uno.WinUI.Lottie" />
+        <PackageReference Include="Uno.WinUI.DevServer" Condition="'$(Configuration)'=='Debug'" />
 
-	<ItemGroup>
-	  <None Remove="Presentation\Components\ItemDetails.xaml" />
-	  <None Remove="Presentation\Components\PageShell.xaml" />
-	  <None Remove="Presentation\HomePage.xaml" />
-	</ItemGroup>
+        <!-- Include all images by default - matches the __WindowsAppSdkDefaultImageIncludes property in the WindowsAppSDK -->
+        <Content Include="Assets\**;**/*.png;**/*.bmp;**/*.jpg;**/*.dds;**/*.tif;**/*.tga;**/*.gif" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder);**\*.svg" />
+        <Page Include="**\*.xaml" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)" />
+        <Compile Update="**\*.xaml.cs">
+          <DependentUpon>%(Filename)</DependentUpon>
+        </Compile>
+        <PRIResource Include="**\*.resw" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)" />
+      </ItemGroup>
+    </Otherwise>
+  </Choose>
 
-	<ItemGroup>
-	  <UpToDateCheckInput Remove="Presentation\Components\ItemDetails - Copy.xaml" />
-	  <UpToDateCheckInput Remove="Presentation\Components\ItemDetails.xaml" />
-	  <UpToDateCheckInput Remove="Presentation\Components\PageShell.xaml" />
-	  <UpToDateCheckInput Remove="Presentation\HomePage.xaml" />
-	  <UpToDateCheckInput Remove="Presentation\Mobile\ItemDetailsPage.xaml" />
-	</ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
+  </ItemGroup>
 
-	<ItemGroup>
-	  <Page Remove="Presentation\HomePage.xaml" />
-	</ItemGroup>
+  <ItemGroup>
+    <UnoImage Include="Assets\**\*.svg" />
+    <EmbeddedResource Include="appsettings.json" />
+    <EmbeddedResource Include="appsettings.*.json" DependentUpon="appsettings.json" />
+    <UpToDateCheckInput Include="**\*.xaml" Exclude="bin\**\*.xaml;obj\**\*.xaml" />
+  </ItemGroup>
 
-	<ItemGroup>
-	  <AdditionalFiles Include="Presentation\Components\HomePage.xaml">
-	    <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
-	    <Generator>MSBuild:Compile</Generator>
-	  </AdditionalFiles>
-	</ItemGroup>
+  <ItemGroup>
+    <Content Remove="Assets\Icons\icon.png" />
+  </ItemGroup>
 
-	<ItemGroup>
-	  <Folder Include="Services\Interop\Store\" />
-	</ItemGroup>
+  <ItemGroup>
+    <None Remove="Presentation\Components\DetailsBox.xaml" />
+    <None Remove="Presentation\Components\ItemDetails.xaml" />
+    <None Remove="Presentation\Components\MobileUser.xaml" />
+    <None Remove="Presentation\Components\PageShell.xaml" />
+    <None Remove="Presentation\Components\StatsBanner.xaml" />
+    <None Remove="Presentation\Components\ToggleIcon.xaml" />
+    <None Remove="Presentation\LoginPage.xaml" />
+    <None Remove="Presentation\MainPage.xaml" />
+    <None Remove="Presentation\NestedPages\ArchivePage.xaml" />
+    <None Remove="Presentation\NestedPages\HomePage.xaml" />
+    <None Remove="Presentation\NestedPages\SettingsPage.xaml" />
+    <None Remove="Presentation\Shell.xaml" />
+  </ItemGroup>
 
-	<ItemGroup>
-	  <AdditionalFiles Remove="Presentation\Components\ItemDetails.xaml" />
-	</ItemGroup>
+  <ItemGroup>
+    <UpToDateCheckInput Remove="Presentation\Components\DetailsBox.xaml" />
+    <UpToDateCheckInput Remove="Presentation\Components\ItemDetails.xaml" />
+    <UpToDateCheckInput Remove="Presentation\Components\MobileUser.xaml" />
+    <UpToDateCheckInput Remove="Presentation\Components\PageShell.xaml" />
+    <UpToDateCheckInput Remove="Presentation\Components\StatsBanner.xaml" />
+    <UpToDateCheckInput Remove="Presentation\Components\ToggleIcon.xaml" />
+    <UpToDateCheckInput Remove="Presentation\LoginPage.xaml" />
+    <UpToDateCheckInput Remove="Presentation\MainPage.xaml" />
+    <UpToDateCheckInput Remove="Presentation\NestedPages\ArchivePage.xaml" />
+    <UpToDateCheckInput Remove="Presentation\NestedPages\HomePage.xaml" />
+    <UpToDateCheckInput Remove="Presentation\NestedPages\SettingsPage.xaml" />
+    <UpToDateCheckInput Remove="Presentation\Shell.xaml" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <UnoImage Include="Assets\Icons\icon.png">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </UnoImage>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Update="Assets\login-background.jpg">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
+    <Content Update="Assets\login-background2.jpg">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
+    <Content Update="Assets\login-background3.jpg">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Page Update="Presentation\NestedPages\ArchivePage.xaml">
+      <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
+      <SubType>Designer</SubType>
+    </Page>
+  </ItemGroup>
+
+  <Import Project="..\winappsdk-workaround.targets" />
 </Project>
