@@ -1,20 +1,18 @@
 using System.Text.Json.Serialization;
 
-namespace ProjectSBS.DataContracts
+namespace ProjectSBS.DataContracts;
+
+public record Currency
 {
-    public record Currency
-    {
-        [JsonPropertyName("date")]
-        public DateTime Date { get; set; }
+    [JsonPropertyName("date")]
+    public DateTime Date { get; set; }
 
-        [JsonPropertyName("amount")]
-        public double Amount { get; set; }
+    [JsonPropertyName("amount")]
+    public double Amount { get; set; }
 
-        [JsonPropertyName("base")]
-        public string BaseCurrency { get; set; }
+    [JsonPropertyName("base")]
+    public string BaseCurrency { get; set; }
 
-        [JsonPropertyName("rates")]
-        public Dictionary<string, double> Rates { get; set; }
-    }
-
+    [JsonPropertyName("rates")]
+    public Dictionary<string, double> Rates { get; set; }
 }
