@@ -1,22 +1,13 @@
-﻿using System.Drawing;
+using System.Drawing;
 
 namespace ProjectSBS.Business.Models;
 
-//public record Tag(
-//    string Name,
-//    Color Color)
-//{
-
-//}
-
-public class Tag 
+public class Tag(
+    int id,
+    string name,
+    Color? color)
 {
-    public string Name { get; set; }
-    public Color Color { get; set; }
-
-    public Tag(string name, Color color)
-    {
-        Name = name;
-        Color = color;
-    }
+    public int Id { get; set; } = id;
+    public string Name { get; set; } = name;
+    public Color? Color { get; set; } = color;
 }
