@@ -1,4 +1,6 @@
-﻿global using CommunityToolkit.Mvvm.ComponentModel;
+global using System.Collections.Immutable;
+global using System.Windows.Input;
+global using CommunityToolkit.Mvvm.ComponentModel;
 global using CommunityToolkit.Mvvm.Input;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Hosting;
@@ -13,24 +15,31 @@ global using ProjectSBS.Business.Models;
 global using ProjectSBS.DataContracts;
 global using ProjectSBS.DataContracts.Serialization;
 global using ProjectSBS.Infrastructure;
-global using ProjectSBS.Presentation;
 global using ProjectSBS.Services.Caching;
 global using ProjectSBS.Services.Endpoints;
-global using Refit;
-global using System.Collections.Immutable;
-global using System.Windows.Input;
-global using Uno.Extensions;
-global using Uno.Extensions.Authentication;
-global using Uno.Extensions.Configuration;
-global using Uno.Extensions.Hosting;
-global using Uno.Extensions.Http;
-global using Uno.Extensions.Localization;
-global using Uno.Extensions.Logging;
-global using Uno.Extensions.Navigation;
-global using Uno.Extensions.Serialization;
+global using Uno.UI;
 global using Windows.ApplicationModel;
 global using Windows.Networking.Connectivity;
 global using Windows.Storage;
-global using ProjectSBS.Presentation.Components;
-global using ProjectSBS.Presentation.Mobile;
 global using ApplicationExecutionState = Windows.ApplicationModel.Activation.ApplicationExecutionState;
+
+global using ProjectSBS.Business;
+global using ProjectSBS.Presentation;
+global using ProjectSBS.Business.ViewModels;
+global using System.Collections.ObjectModel;
+global using System.Globalization;
+global using ProjectSBS.Services.Analytics;
+global using ProjectSBS.Services.Interop;
+global using ProjectSBS.Services.Items;
+global using ProjectSBS.Services.Items.Tags;
+global using ProjectSBS.Services.Items.Billing;
+global using ProjectSBS.Services.Items.Filtering;
+global using ProjectSBS.Services.Navigation;
+global using ProjectSBS.Services.Notifications;
+global using ProjectSBS.Services.Settings;
+global using ProjectSBS.Services.Storage;
+global using ProjectSBS.Services.Storage.Data;
+global using ProjectSBS.Services.User;
+global using ProjectSBS.Services.Dialogs;
+global using CommunityToolkit.Mvvm.Messaging;
+global using ProjectSBS.Presentation.NestedPages;
