@@ -1,4 +1,4 @@
-﻿using Microsoft.AppCenter.Crashes;
+using Microsoft.AppCenter.Crashes;
 using Service = Microsoft.AppCenter.Analytics;
 using Microsoft.AppCenter;
 
@@ -8,19 +8,16 @@ public class AnalyticsService : IAnalyticsService
 {
     public AnalyticsService()
     {
+#if !HAS_UNO
         AppCenter.Start("ea15cf3d-e346-483c-b0b3-618fa515fd50",
                   typeof(Service.Analytics), typeof(Crashes));
-
-        Dictionary<string, string> appLaunchSettings = new()
-            {
-                { "123", $"Hello there" }
-            };
-
-        TrackEvent("App launched", appLaunchSettings);
+#endif
     }
 
-    public void TrackEvent(string eventName, IDictionary<string, string> properties = null)
+    public void TrackEvent(string eventName, IDictionary<string, string>? properties = null)
     {
+#if !HAS_UNO
         Service.Analytics.TrackEvent(eventName, properties);
+#endif
     }
 }
