@@ -1,18 +1,16 @@
-﻿namespace ProjectSBS.Services.Items.Filtering;
+namespace ProjectSBS.Services.Items.Filtering;
 
 public class ItemFilterService : IItemFilterService
 {
-    public List<FilterCategory> Categories { get; }
+    public List<Tag> Categories { get; }
+    public Tag SelectedCategory { get; set; }
 
-    public ItemFilterService(IStringLocalizer localizer)
+    public ItemFilterService(IStringLocalizer localizer, ITagService tags)
     {
-        //TODO: Add proper Selectors for FilterCategories
-        Categories = new()
-        {
-            new(localizer["Home"], "\uE80F"),
-            new(localizer["Upcoming"], "\uE752", i => !i.IsPaid),
-            new(localizer["Overdue"], "\uEC92", i => i.Item.Name is "Sample Item 1"),
-            new(localizer["Expensive"], "🥩", i => i.Item.Billing.BasePrice > 50),
-        };
+        Categories =
+        [
+            new Tag(-1, localizer["All"], null), .. tags.Tags
+        ];
+        SelectedCategory = Categories[0];
     }
 }
