@@ -1,9 +1,14 @@
-﻿namespace ProjectSBS.Services.Items;
+namespace ProjectSBS.Services.Items;
 
 public interface IItemService
 {
+    event EventHandler<IEnumerable<ItemViewModel>> OnItemsInitialized;
+    event EventHandler<IEnumerable<ItemViewModel>> OnItemsChanged;
     Task InitializeAsync();
     IEnumerable<ItemViewModel> GetItems(Func<ItemViewModel, bool>? selector = null);
+    bool ClearItems();
     void NewItem(Item item, List<ItemLog>? logs = null);
-    ItemViewModel ScheduleBilling(ItemViewModel item, List<ItemLog> logs);
+    void UpdateItem(ItemViewModel item);
+    void ArchiveItem(ItemViewModel item);
+    void DeleteItem(ItemViewModel item);
 }
