@@ -1,35 +1,32 @@
-﻿using ProjectSBS.Services.Items.Billing;
-using ProjectSBS.Services.Notifications;
-using ProjectSBS.Services.Storage.Data;
-
 namespace ProjectSBS.Services.Items;
 
-public class ItemService : IItemService
+public class ItemService(IDataService dataService) : IItemService
 {
-    private readonly IBillingService _billing;
-    private readonly IDataService _dataService;
-    private readonly INotificationService _notification;
+    private readonly IDataService _dataService = dataService;
 
-    private readonly List<ItemViewModel> _items = new();
+    private bool _isInitialized;
 
-    public ItemService(
-        IBillingService billing,
-        IDataService dataService,
-        INotificationService notification)
-    {
-        _billing = billing;
-        _dataService = dataService;
-        _notification = notification;
-    }
+    private readonly List<ItemViewModel> _items = [];
+
+    public event EventHandler<IEnumerable<ItemViewModel>>? OnItemsInitialized;
+    public event EventHandler<IEnumerable<ItemViewModel>>? OnItemsChanged;
 
     public async Task InitializeAsync()
     {
-        var (items, logs) = await _dataService.InitializeDatabaseAsync();
+        if (_isInitialized)
+        {
+            return;
+        }
+
+        var (items, _) = await _dataService.InitializeDatabaseAsync();
 
         foreach (var item in items)
         {
-            AddNewItem(item, logs);
+            AddNewItem(item);
         }
+
+        OnItemsInitialized?.Invoke(this, _items);
+        _isInitialized = true;
     }
 
     public IEnumerable<ItemViewModel> GetItems(Func<ItemViewModel, bool>? selector = null)
@@ -42,41 +39,76 @@ public class ItemService : IItemService
         return _items.Where(selector);
     }
 
+    public bool ClearItems()
+    {
+        try
+        {
+            _items.Clear();
+            OnItemsChanged?.Invoke(this, _items);
+            _isInitialized = false;
+            return true;
+        }
+        catch
+        {
+            return false;
+        }
+    }
+
     public void NewItem(Item item, List<ItemLog>? logs = null)
     {
-        AddNewItem(item, logs ?? new());
+        AddNewItem(item, logs ?? []);
 
-        var itemsList = _items
-            .Select(itemViewModel => itemViewModel.Item)
-            .ToList() ?? new();
-        _dataService.SaveDataAsync(itemsList);
+        SaveDataAsync();
     }
 
     private void AddNewItem(Item item, List<ItemLog>? logs = null)
     {
         ItemViewModel itemVM = new(item);
 
-        logs ??= new();
-
-        itemVM.Initialize(logs);
-
-        ScheduleBilling(itemVM, logs);
         _items.Add(itemVM);
     }
 
-    public ItemViewModel ScheduleBilling(ItemViewModel itemVM, List<ItemLog> logs)
+    public void UpdateItem(ItemViewModel item)
     {
-        var item = itemVM.Item;
-        var paymentDates = _billing.GetFuturePayments(item.Billing.InitialDate, item.Billing.PeriodType, item.Billing.RecurEvery);
+        var index = _items.IndexOf(item);
+        _items[index] = item;
 
-        Task.Run(() =>
+        SaveDataAsync();
+        item.Updated();
+    }
+
+    public void ArchiveItem(ItemViewModel item)
+    {
+        if (item?.Item is not { } i)
         {
-            foreach (var date in paymentDates)
-            {
-                _notification.ScheduleNotification(item.Id, item.Name, DateTime.Now.ToString(), date, new TimeOnly(8, 00));
-            }
-        });
+            return;   
+        }
 
-        return itemVM;
+        i.Status.Add(new(item.IsArchived ? State.Active : State.Archived, DateTime.Now));
+
+        SaveDataAsync();
+        item.Updated();
+    }
+
+    public void DeleteItem(ItemViewModel item)
+    {
+        _items.Remove(item);
+
+        SaveDataAsync();
+        item.Updated();
+    }
+
+    private void SaveDataAsync()
+    {
+        var itemsList = _items
+            .Select(itemViewModel => itemViewModel.Item)
+            .ToList() ?? [];
+
+        if (itemsList is { })
+        {
+            _dataService.SaveDataAsync(itemsList);
+        }
+
+        OnItemsChanged?.Invoke(this, _items);
     }
 }
