@@ -1,8 +1,17 @@
-﻿namespace ProjectSBS.Services.User;
+namespace ProjectSBS.Services.User;
 
 public interface IUserService
 {
-    Task<Business.Models.User?> GetUser();
+    bool IsLoggedIn { get; }
+    bool NeedsRefresh { get; }
+
+    event EventHandler<Business.Models.User?>? OnLoggedInChanged;
+
+    Task<bool> AuthenticateAsync(bool silentOnly = false);
+    Task<Business.Models.User?> RetrieveUser();
+
     Task<bool> UploadData(string content, string relativeLocalPath, CancellationToken token = default);
     Task<Stream?> RetrieveData(string relativeLocalPath, CancellationToken token);
+
+    void Logout();
 }
